Use cv::Mahalonobis() for distance measure. This required more complicated binding, so boost::bind was used. This required a switch to boost::shared_ptr instead of C++0x's std::shared_ptr.
The -std=c++0x flag was removed from the makefile since std::shared_ptr was
the only reason it was needed.  boost libraries should be used from now on
instead of c++0x standard features. (This is better because who know's what
gcc will do for c++0x support between now and finalization of the standard)



git-svn-id: http://benbelly.unfuddle.com/svn/benbelly_ms/epenthesis@101 c3643d77-5941-4ca3-9a90-978904f1a040